+++ examples/my_line_rules.py
@@ -39,6 +39,8 @@ class SpecialChars(LineRule):
                                "Comma separated list of characters that should not occur in the title")]
 
     def validate(self, line, _commit):
+        self.log.debug("SpecialChars: This line will be visible when running `gitlint --debug`")
+
         violations = []
         # options can be accessed by looking them up by their name in self.options
         for char in self.options['special-chars'].value: