+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.14.0 (TBD) ##
 - [User-defined Configuration Rules](TODO) allow users to dynamically change gitlint's configuration and/or the commit *before* any other rules are applied.
+- Users can now use `self.log.debug("my message")` for debugging purposes in their user-defined rules
 - Breaking: User-defined rule id's can no longer start with 'I', as those are reserved for built-in gitlint ignore rules.
 
 ## v0.13.1 (2020-02-26)